requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
google-generativeai==0.3.0
PyPDF2==3.0.1
python-docx==1.1.0
bcrypt==4.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
mysql-connector-python==8.1.0